mythwelcome: rename commandlineparser.* to mythwelcome_…
